+++ lib/jackrabbit.js
@@ -1,132 +1,97 @@
-var amqp = require('amqplib');
-var EventEmitter = require('events').EventEmitter;
+var amqp = require('amqplib/callback_api');
 
-var Queue = require('./queue');
+// All brokers have a default exchange ('')
 
-function JackRabbit(url, prefetch) {
+// Creating a channel named '' and exclusive: true
+// creates a private channel that is named by the amqp server.
+// http://www.squaremobius.net/amqp.node/doc/channel_api.html
+
+// Channels are just 'light connections'
+// https://www.rabbitmq.com/tutorials/amqp-concepts.html
+
+module.exports = Jackrabbit;
+
+function Jackrabbit(url, overrides) {
   if (!url) throw new Error('url required for jackrabbit connection');
 
-  EventEmitter.call(this);
+  // Extend default options with overrides
+  options = _.extend({
+    exchange: '',
+    prefetch: 1
+  }, overrides);
 
-  this.connection = null;
-  this.channel = null;
-  this.prefetch = prefetch || 1;
-  this.queues = {};
+  var connection, channel, queues;
+  var queues = {};
+  var handlers = [];
 
-  amqp
-    .connect(url)
-    .then(this.createChannel.bind(this))
-    .then(this.onChannel.bind(this))
-    .catch(this.onConnectionErr.bind(this));
-}
+  // Expose public methods
+  var rabbit = {
+    handle: handle,
+    publish: publish,
+    close: close
+  };
 
-module.exports = function createJackRabbit(url, prefetch) {
-  return new JackRabbit(url, prefetch);
-};
+  EventEmitter.call(rabbit);
+  amqp.connect(url, onConnection);
+  return rabbit;
 
-JackRabbit.prototype = Object.create(EventEmitter.prototype);
+  // handle messages sent to queue 'queueName'
+  function handle(queueName, handler, overrides) {
+    assertQueue(queueName, options, onQueue);
 
-JackRabbit.prototype.createChannel = function(connection) {
-  this.connection = connection;
-  this.connection.once('close', this.onClose.bind(this));
-  return connection.createChannel();
-};
-
-JackRabbit.prototype.onChannel = function(channel) {
-  this.channel = channel;
-
-  // Create a reply queue
-  this.channel.replyHandlers = {};
-  this.channel.assertQueue('', { exclusive: true })
-    .then(function(replyTo) {
-      this.channel.replyName = replyTo.queue;
-      this.channel.consume(this.replyName, this.onReply.bind(this), { noAck: true });
-      this.emit('connected');
-    }.bind(this));
-
-  this.channel.on('close', this.onChannelClose.bind(this));
-  this.channel.prefetch(this.prefetch);
-};
-
-JackRabbit.prototype.onReply = function(msg) {
-  var id = msg.properties.correlationId;
-  var replyHandler = this.channel.replyHandlers[id];
-  if (!replyHandler) return;
-
-  var body = msg.content.toString();
-  var obj = JSON.parse(body);
-  replyHandler(null, obj);
-};
-
-JackRabbit.prototype.close = function() {
-  this.connection.close();
-};
-
-JackRabbit.prototype.onClose = function() {
-  this.emit('disconnected');
-};
-
-JackRabbit.prototype.onConnectionErr = function(err) {
-  this.emit('disconnected', err);
-};
-
-JackRabbit.prototype.onChannelClose = function() {
-  this.emit('disconnected');
-};
-
-JackRabbit.prototype.queue = function(name, options) {
-  return new Queue(this.channel, name, options);
-};
-
-JackRabbit.prototype.create = function(name, options, done) {
-  if (!done) {
-    if (typeof options === 'function') {
-      done = options;
-      options = {};
-    }
-    else {
-      done = function() {};
+    function onQueue(err, queue) {
+      if (err) return bail(err);
+      channel.consume(queueName, handler, options);
     }
   }
-  var queue = new Queue(this.channel, name, options);
-  queue
-    .once('ready', function onQueueReady(info) {
-      this.queues[name] = queue;
-      done(null, queue, info);
-    }.bind(this))
-    .once('error', function onQueueErr(err) {
-      done(err);
-    });
-};
 
-JackRabbit.prototype.destroy = function(name, done) {
-  this.channel
-    .deleteQueue(name)
-    .then(onSuccess)
-    .catch(onFail);
+  // publish messages to the queue 'queueName'
+  function publish(queueName, data, overrides) {
+    // first, assert that the queue exists
+    if (queues[queueName]) onQueue(undefined, queues[queueName]);
+    else channel.assertQueue(queueName, options, onQueue);
+  }
 
-  function onSuccess() { done(null, true); }
-  function onFail(err) { done(err); }
-};
+  function close() {
+    if (connection) connection.close();
+  }
 
-JackRabbit.prototype.purge = function(name, done) {
-  this.channel
-    .purgeQueue(name)
-    .then(onSuccess)
-    .catch(onFail);
+  function assertQueue(name, overrides, done) {
+    if (queues[queueName]) {
+      return onQueue(undefined, queues[queueName]);
+    }
+    var options = _.extend({
+      exclusive: false,
+      durable: true,
+      messageTtl: undefined,
+      maxLength: undefined
+    }, overrides);
 
-  function onSuccess(response) { done(null, response.messageCount); }
-  function onFail(err) { done(err); }
-};
+    channel.assertQueue(queueName, options, onQueue);
 
-JackRabbit.prototype.publish = function(name, obj, replyHandler) {
-  this.queues[name].publish(obj, replyHandler);
-};
+    function onQueue(err) {
+      if (err) return done(err);
+      queues[queueName] = true;
+    }
+  }
 
-JackRabbit.prototype.handle = function(name, handler) {
-  this.queues[name].subscribe(handler);
-};
+  function bail(err) {
+    connection = undefined;
+    channel = undefined;
+    rabbit.emit('error', err);
+  }
 
-JackRabbit.prototype.ignore = function(name) {
-  this.queues[name].unsubscribe();
-};
+  function onConnection(err, conn) {
+    if (err) return bail(err);
+    connection = conn;
+    connection.on('close', bail.bind(new Error('connection closed')));
+    connection.createChannel(onChannel);
+  }
+
+  function onChannel(err, chan) {
+    if (err) return bail(err);
+    channel = chan;
+    channel.on('close', bail.bind(new Error('channel closed')));
+    channel.prefetch(options.prefetch);
+  }
+}
